Removed the parameter validation introduced in aggregation and weights
Moved the additional parameter validation added in the functions in
weights.p and aggregation.py modules, instead relying on the user to use
the function correctly based on the documentation (we're all adults
here). Changed behaviour based on Python's preference for EAFP parameter
validation instead of LBYL.

Decided only to keep the _handle_axis function to transform axis
arguments given in string format to ints - put this in _validation.py
module.